api_app/requirements.txt
psycopg2==2.7.7
flask==1.1.1
uwsgi==2.0.18
testing.postgresql==1.3.0